fixed activity title to be left-aligned
