+++ screens/AllActivitiesScreen.js
@@ -2,7 +2,7 @@
 import React, { useContext,useEffect } from "react";
 import { View, Text,Button } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
-import { ActivityContext } from '../components/ActivityList'; // Import your ActivityContext
+import { ActivityContext } from '../components/ActivityContext'; 
 import { Entypo } from '@expo/vector-icons';
 import { COMMON_STYLES } from "../components/styles";
