+++ screens/SpecialActivitiesScreen.js
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect } from 'react';
 import { View,Text, Button} from 'react-native';
 import { useNavigation } from '@react-navigation/native';
-import { ActivityContext } from '../components/ActivityList'; 
+import { ActivityContext } from '../components/ActivityContext'; 
 import { COMMON_STYLES } from "../components/styles";
 import { Entypo } from '@expo/vector-icons';
